Push all of the user-configurable options out of rawenv, and into the database. This keeps us from having to spam rawenv with new config options, plus it gives us tighter control over what can be easily configured.
All of the config options can be set and viewed using the various
``tc config'' commands.  Each one allows for their particular
``configlet'' to get viewed and modified.

Additionally, a config command called __DSVERSION__ was added to track
database schema and config data changes.  This will be used for the
forthcoming setup and migration script.

TODO: Add support to the jails and ports_trees table for customizable mounts.