Repocopy all Perl modules into a Tinderbox subdirectory.  This gives us better namespace protection.
All modules must now be preceeded by Tinderbox::.  TinderboxDS has been
updated in such a way to make this change transparent to developers using
the existing APIs.  However, a new internal API call has been added for
convenience.  To test if an object is of a certain type, use the _isA()
static method in TinderboxDS.  For example, to test whether a given object
is a Jail, use the following:

if (_isA("Jail", $object)) {
    ...
}