Introduce a more extensible way of creating Jail source trees and PortsTrees.  Defaults to CVSUP, but now has the capability for CSUP, USER, and NONE.
In the CVSUP/CSUP/USER cases, an appropriate update.sh script is
now created so that updating the trees later on simply becomes a
question of the code calling {jail,portstrees}/<name>/update.sh into
a logfile into the same directory, removing a lot of special-case
code elsewhere.

XXX placeholders have been put in the README for future documentation
and explanation of this feature.