Add a new configuration section for logging.  One can now specify a log directory and whether or not to copy log files to this directory, or have them symlinked (the default).  The idea here is to allow one to centralized all Tinderbox logs into one directory.  For example, one could create /space/tblogs, then do:
tc configLog -d /space/tblogs

After each build, a directory will appear under /space/tblogs as
BUILD-DATE (e.g. 7.2-FreeBSD-20091012135501).  Within this directory, all
relevant log files for the portbuild will be found.  This includes the
tinderbuild log (if not using tinderd), the make logs, and the individual
port build logs.  If log copying is enabled, the logs will be copied to
this directory instead of just using symlinks.

Official documentation, and a relatively quick MFH to follow.  This is the
last major feature before 3.3 is released.

Requested by:	itetcu